// model/SearchListItem.ts
export type SearchListItem = {
  id: number;
  site: string;
  channel: string;
  currentState: string;
  searchKeyword: string;
  searchStartDate: string;
  searchEndDate: string;
  workerIp: string;
  startDate: string;
  endDate: string;
  createDate: string;
  count: number;
  s3Url: string;
  fileSize: number;
  memId: number;
};